Remove builtin padding from .ve-ce-documentNode
I think it's a lot easier if setting up paddings and margins on the
DocumentNode that work well with everything else on the page is the
responsibility of the Target, rather than DocumentNode/Surface.

For example, we set .ve-ce-documentNode padding in:
* ve.init.sa.css (standalone Target in the demos)
* ve.init.mw.DesktopArticleTarget-vector.css (VE-MW)
* ve.ui.TargetWidget.css (used in VE-MW dialogs)

Also, follow-up to d426f806639fd1682be8eef241a1cd31172d8f91:
removing a dead line of CSS (it was always overridden), and
a misleading comment in TargetWidget.

Bug: T162493
Change-Id: Ie2431ea9816e563702a84c202d7e52e4f1ca6e85