rebaser: Hacky implementation display names
Allow user to change their display name, and display this name
next to their selection and in a list in a sidebar.

Adds handling of author disconnects and session usurpation:

When clients register, they receive a token which they store in sessionStorage,
then send back to the server when reconnecting. The server validates this token
and allows the client to resume the session.

Also put all author data in one map.

Change-Id: I9e6f1b01ac7cc55d532b7b75494324cdcb99c60c