Update patch set 1
Patch Set 1:

However, regardles of done() or then(), it is deprecated as of this release, so nicer would be to use the lifecycle promises instead, like was done for prompt() and alert(). But if not sure which to use, then changing done() to then() will work for now (with deprecation warning).

Patch-set: 1
Reviewer: Krinkle <34@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0