ui.FormatAction: fixup selection for empty conversions
Converting an empty node needs a small selection fixup afterwards, otherwise
the selection will be displayed outside the new empty node. This causes issues
with the display of the current format in the toolbar, and with hitting enter
if no content is entered. Don't always reapply the selection, because the
automatic behavior is better if isolateAndUnwrap has actually acted.

Bug: T151594
Change-Id: Ibf143e7954919d0d9efacd53205d5451461a3115