Update patch set 1
Patch Set 1:

> Don't we normally have config like this in the positive direction,
 > i.e. [showDir=true] rather than [hideDir]?

Well, this very widget has "hideCodeInput".  And "requireDir" hides the "auto" option if it is true, so it is similar to "hideDir" hiding the directionality entirely if true.

But mostly I didn't want to break backward compat.  If you weren't afraid of breaking eggs, 'requireDir' and 'hideDir' should be combined somehow into a three-value option: dir=no/no-auto/yes or something like that.

Patch-set: 1
Label: Verified=0