Update patch set 2
Patch Set 2: Code-Review+1

Note that for VE/ve-mw, you can safely do this in two steps if needed given that beforeEach/afterEach already exist in 1.x as well. So you can migrate first, and then upgrade. (So that you're not blocked on upgrading ve-mw at the same time as the next pull through, it can be done earlier in ve-standalone and later in ve-mw. Otherway around wouldn't work as ve-mw will fail with non-migrated ve-standalone tests).

Patch-set: 2