Update patch set 3
Patch Set 3:

This isn't just an IE problem, other browsers also reject multi-root input, although they fail differently. Chrome, for example, doesn't throw an exception but instead returns a document with a <parsererror> tag and the first root.

While this hack looks like it'll work, I don't like it that much, and I'd like to know what code is passing multi-root input to this function and why. Technically it's the caller's fault since the docs say the input has to be valid XML.

Patch-set: 3