Update patch set 12
Patch Set 12: Code-Review-1

In testing in VE-MW, when I cut a multi-paragraph selection including a bare reference, checkExpected() threw, because (I think) the full attributes aren't preserved.

Specifically, "actual" is missing the "mw", "originalMW", and "contentsUsed" key/values. Is this something that needs fixing in VE-MW/Cite to adapt to this (in which case it's a breaking change, and probably other things might need similar adaption before this is merged) or is this patch at fault?

Patch-set: 12
Label: Code-Review=-1