Update patch set 1
Patch Set 1:

I based https://gerrit.wikimedia.org/r/377237/ on this (conceptually, not in patchset terms).

I kept the pattern of having the document object do the notifying, but using the actual DM node objects instead of their linmod data representations. There are a few reasons this is necessary: one is that a transaction that ostensibly preserves a node can actually end up rebuilding it afresh in the current implementation. It could also be useful to have a handle on the actual node object for various purposes. The notifications happen immediately when TreeModifier attaches/detaches a node - this is now reasonable because TreeModifier keeps the tree consistent after each atomic step of transaction processing.

Patch-set: 1
Reviewer: Divec <839@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=0