LinearEnterKeyDownHandler: make sure the range is in a cursorable location
If the range is such that we try to act on the document node, everything will
break because it can't be closed. As such, make sure we're in the closest
possible cursorable position.

In my testing this seems to come up in interactions with block content which
has a paragraph following it, where the selection winds up focused between the
block and the paragraph instead of inside the paragraph. In practice this code
thus mostly works as `offset = offset + 1`.

Bug: T148679
Change-Id: I00b900646b3d335bfd3c74f8db1b32cf3cdf429d