Update patch set 1
Patch Set 1:

I can't reproduce the bug as described, but it sounds like this is fixing the symptom rather than the cause. The cause here is the cursor being placed in an invalid position to begin with. Linear selections should either be in text (a ContentBranchNode), or spanning a FocusableNode. If we fix that then this patch is probably not necessary.

Patch-set: 1
Reviewer: Esanders <658@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=0