ce.Surface: handle table-adjacent observed selection changes better
Moving to the closest cursorable location produces better results, and means
clicking to the left of a table no longer vanishes the cursor.

There's a remaining bug when a table is the very first node in a document, and
you click to the left of it *twice*, which deselects the block slug and places
the cursor in an invalid position. This needs more changes to various
optimizations selection-changes try to make to avoid, however.

Bug: T148679
Bug: T176936
Change-Id: I8a3dde56d52342316c00a909ac9369b2a9f6b409