Scroll to the next highlight after find and replace has replaced text
Right now when you're doing find and replace, each time you replace, the
next occurrence of the term is highlighted, but not scrolled to. This is
unexpected, and can lead to you being quite far away from the highlighted
text if you keep replacing things. This patch fixes that by scrolling to
the highlighted text each time something is replaced.

Bug: T177514
Change-Id: Ic5a39f681260c910668bb1ea3a27aa2c62f14383