Update patch set 2
Patch Set 2:

> Not end-user configurable, but configurable by people writing VE
 > implementations.

That might be nice for the developers as it would allow them enable/disable it if they want. I don't think it would be good for the end-user because it could spoil "consistency". They wouldn't expect to different incarnations of the visual editor in different places. If it did, they might find it difficult to remember "which" incarnation bahaves in which way. So, it's better to not allow configuring the "tab" behaviour.

 > Also I agree somewhat with this post: https://www.mediawiki.org/w/index.php?title=Topic:Tzmpo1gjj2gv2ugy&topic_showPostId=u3wmlatd6fumiz58#flow-post-u3wmlatd6fumiz58
 > 
 > Tab only leaving the editor in certain contexts could be confusing.

This makes sense. So, it's better to abandon this for now. The issue specified in the commit message should better be resolved by a keyboard shortcut instead.

Patch-set: 2
Reviewer: Kaartic <3683@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0