Update patch set 2
Patch Set 2:

'A &nbsp; B' is likely to mean the source intended there to be three spaces, so we leave it as is.

In '2&nbsp;kgs' it is impossible to tell if the nbsp was added by the user or the browser, so we have to pick one camp.

Patch-set: 2
Reviewer: Esanders <658@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0