Update patch set 2
Patch Set 2: Code-Review-1

You can do this more simply by checking !this.initialized in onNewChange.

However, for a proper solution we'll need to fix the concurrency in the server. If a newChange event arrives before an initDoc event, that's the server's fault for sending them in that order. Depending on how badly the server screwed up, the change delivered in the newChange event may also be part of the initDoc series of changes, or it may not be (or for multiple newChange events, a mix of these can occur).

A correct workaround on the client side would involve buffering premature newChange events and flushing that buffer when initialization is complete, discarding any duplicate changes. A correct fix on the server would probably be quite tricky since we're using socket.io namespaces for broadcasting right now.

Patch-set: 2
Reviewer: Catrope <5@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=-1