docs: Convert multi-line comments to /**/ syntax
The problem I see with using //s for multi-line comments is that written
comments starting with '//'s are meant to be short and 'snappy'. Endless
rows of // comments result in a mess.

If one is expanding a comment to a second line, generally because either
(a) one is expanding the functionality, or (b) one has detected a subtle
issue with the code that was not apparent on first writing, and which is
now being resolved, one should step back and write a proper, full-length
comment, not just add a few short words on a new line to explain the too
complication situation.

Change-Id: Ia813c83b0e71aebb1489e9efd35762898f6498b0