Update patch set 2
Patch Set 2:

This is going to be a pain to pull through to VE-ME because the Cite extension calls the old methods and so sets up a circular dependency. Leave the old method names in as deprecated aliases for a day so we can pull through cleanly?

Patch-set: 2
Reviewer: Jforrester <366@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=0