Update patch set 1
Patch Set 1:

I kind of feel that the zero-width space thing is a hack, and something related to this should be rolled into OOUI's labelElement -- let a labelElement always have the CSS classes flagging it as such, and have a separate class to distinguish empty/present labels. That'd be a bit of a backwards-compatibility change, though.

Patch-set: 1