Update patch set 5
Patch Set 5: Code-Review+2

It's actually `42px` in height, which is sufficient (we long for minimum 40 for touch devices). 
This looks fine for now. We could revisit it with let's say 16px, but then you have the horizontal alignment with the menu handler broken. 
I'd also not be surprised if the checkmarks come back here (for different reasons) ;P

Patch-set: 5
Reviewer: VolkerE <2604@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=+2