PlainTextStringTransferHandler: Allow as paste
ve.ce.Surface:2365 assumes that plain text can be pasted into a Surface, and
tries to do so if the pasted content turned out to be entirely plain text.
This always failed, as it was actually forbidden. This never worked, and
actually enabling it breaks many assumptions, mostly to do with annotation
coverage.

The main visible effect of allowing paste of plain text is that pasting
wikitext from a different NWE instance will no longer go through a convert-to-
source phase which adds <nowiki> tags. This is because NWE overrides onCopy,
to explicitly set the selection contents as text/plain.

A potential side effect is that people pasting plain text into NWE that
contains wikitext-like sequences which are not valid wikitext will be
surprised by errors. I suspect this is less common.

Bug: T190590
Change-Id: I2db85c6ed3cdc14cd8d5ce77a735a216424a6059