Update patch set 1
Patch Set 1: Code-Review-1

Action is `remove` if you press the Remove button at the bottom of the inspector.

This patch just seems to make the cancel button continue its current bad-behavior of not actually discarding the change in the inspector.

I think we actually need an extra branch in that if statement, so it can handle action==undefined separately from the remove case.

Patch-set: 1
Label: Code-Review=-1