Improve behavior of newlines and non-text content elements in find-and-replace
"Find and replace" can now match and modify the following within
document data:
* literal newlines (*not* paragraph breaks, only actual newline
  characters)
* non-text elements which act as content (e.g. comments `<!-- -->`
  or inline images), represented using 2x U+FFFC characters

Documentation notes:

When replacing, you can't preserve non-text elements using capturing
groups (they become empty strings; this seems better than inserting
U+FFFC). They mostly exist in the searched text to allow anchors
like `^`, `$`, `\b` etc. to behave "sanely".

To ensure that we never try to select or replace half of a node, if
a search would match only the opening element, is it extended by one
to also match the closing; if a search would match only the closing
element, this match is skipped and the search is continued from the
next position.

I am not entirely happy with this behavior; because each node is
represented by two replacement characters in the text version, we
have to handle "partial" matches somehow and it will be awkward no
matter how we do it.

I think the ideal solution would be to transform the input regexp so
that it can never match an U+FFFC (removing it from explicit ranges,
replacing `\uFFFC` with `[^\s\S]`, replacing `.` with `[^\n\uFFFC]`,
and so on for other tokens like `\S`), and then introduce a special
syntax to match a node, which is replaced with `\uFFFC{2}`. This seems
doable, but I don't think anyone has done it before, so we would have
to write a bunch of code to parse and transform regexps, or at least
include a library like XRegExp and write a plugin for it. Maybe
someone will implement this in the future.

Bug: T188651
Bug: T184163
Change-Id: Ib93b3de4039eb95a2aacd52a379616b7f7e4c88d