Update patch set 2
Patch Set 2: Code-Review-1

I just realized that your original fix also drastically improved the performance of plaintext pastes in NWE (T189556) so I would be sad if we had to revert it.

Is it possible to instead "unregister" this handler for VE-MW only? We already unregister data transfer handles for NWE (see code at the bottom of ve.ui.MWWikitextDataTransferHandlerFactory.js).

Patch-set: 2