Update patch set 1
Patch Set 1:

I think this is correct. selectionBefore is the most recent selection before the start of the last transaction stack - so when you start a new staging context the stack is empty and the last transaction stack is still the stack before this one, and so selectionBefore is the selectionBefore value from that stack.

Have added comments to clarify, and put in the correct value for when multiple staging stacks are used (which isn't often).

Patch-set: 1
Label: Verified=0