Update patch set 6
Patch Set 6: Code-Review+1

I noticed in local testing that without this patch, trying to select the inner contents of a link and typing results in plain text on demos/desktop (link is lost), whereas on demos/mobile it somehow preserved the empty link and put the text into it.

With the patch applied, demos/desktop behaves the same as demos/mobile which is nice. And either way, the extra Label field is very nice. I've missed this in VE when comparing editing to other software (Google Docs, Apple Pages, Word, ..), which all seem to do the label thing as well.

Naturally, because VE allows italic/bold annotations and such, any of those that don't apply to the entire label previously will be lost, but that seems fine and not unexpected from UX, should be rare anyhow. Very nice! And look, it still rebases cleanly after 2 years!

Patch-set: 6