Linear/TableKeyDownHandlers: account for table captions
`Tab` in a table caption should move out of the caption, rather than the
cursor being trapped there. `Up` from the first row of a table should move
into the caption rather than skipping it.

Also, `shift+tab` in the first cell of a table when wrap is enabled can be
treated as pressing `up`. (Otherwise the caption can be tabbed out of, but not
into, despite looking like part of the table.)

Bug: T192163
Change-Id: Icd229cef9acb5e2ac70c6ddbff5b12a193f706fa