Update patch set 2
Patch Set 2: Code-Review+2

I had to put in some effort to understand what this is fixing. For example, place the cursor in the middle of a link annotation, and choose Insert → Comment from the menu. The label of the primary action of the inspector should be "Insert" and not "Done". I'm not sure if there were any worse consequences of using the wrong mode, the comment inspector at least seemed to work just fine.

Patch-set: 2
Reviewer: Bartosz Dziewoński <417@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=+2