Update patch set 2
Patch Set 2:

> Shouldn't up/down in a table caption behave the same as
 > tab/shift-tab?

My logic for not adding that was that the caption is a multiline context, so up/down might be intended as navigation within the caption. Same as how when the focus is actually within a table cell, up/down doesn't switch cells.

Patch-set: 2
Reviewer: DLynch <2915@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0