[WIP] FocusableNode: Allow the highlights to have holes (using SVG)
Using the PolyBool library, we compute a polygon describing the
FocusableNode highlights and put it into a single SVG <path> node,
instead of using multiple <div>s.

This is meant to allow for:
* Having "holes" inside the highlights, e.g. for an image caption
* Preventing highlights from overlapping out-of-flow elements
  (since we can now trim off that part)

Since FocusableNode's $focusable can now contain editable content (not
covered by the highlight), tweak some event handling to fix highlight
rendering when interacting with it.

As there is now just one ve-ce-focusableNode-highlight per
FocusableNode, we can set opacity directly on it (previously this
wouldn't work well if multiple highlights overlapped), which lets us
remove an Internet Explorer workaround.

Since we use a <path> element now, the AlienNode fancy background
pattern has to be applied with 'fill' rather than 'background-image',
and that doesn't accept CSS3 background values like
'repeating-linear-gradient()'. Recreate it as a SVG <pattern>.

Shout-out to the flight W6 1176 and the train PKPIC 83174 for their
contributions to this patch.

Bug: T53202
Bug: T190226
Change-Id: I9570422e2c8850f3d468eea4d213db5f785fba3c