ve.ce.FocusableNode: Remove "ve-ce-focusableNode-highlight-relocatable-marker"
Originally implemented in 2013 in e71655304eba7521e036946e295923b3f60d6145
to "achieve a native drag marker". It doesn't seem to be needed anymore.
If anything, the drag marker looks worse with this <img> than without it.

This lets us remove unreliable code trying to hide this marker on
right-click to avoid the browser displaying menu options to copy/save
the 1px GIF image (it only works on Firefox for me, not Chrome/Edge).

Instead we need to set dummy data transfer data for Firefox to allow
users to drag elements other than <img>.

Change-Id: I287f3dca57406390e09443a19b03e2ab68741632