Update patch set 2
Patch Set 2:

Haven't tested, but would be awesome if we can remove this. This was also a source of performance problems at some point in one or two major browsers due to them synchronously trying to set up the image (internal "network" fetch, decode base64, parse binary, headless rendering of the image) - all synchronously in the particular browsers that it affected (somehow unlike normal urls, these browsers did the whole thing synchronously as part of setting the "src" attribute in the case of data uris).

Patch-set: 2