Update patch set 3
Patch Set 3: Code-Review-1

This does have a visible side-effect of making the icons in comments bigger, like so https://www.dropbox.com/s/ar4qgnqbaz9e4u3/Screenshot%202018-05-30%2010.39.14.png?dl=0 (pre-patch on left, post-patch on right). This makes them not blend in with normal line-height text.

It also slightly increases the amount of text included in the excerpt, just by virtue of dumping the whole comment in there and letting CSS trim it, but that seems fine since setLabel will just be using it as text. Performance implications with excessively huge comments maybe? But that's a niche case.

Patch-set: 3
Label: Code-Review=-1