Update patch set 2
Patch Set 2:

Perhaps this is too much paranoia, but maybe we shouldn't allow HTML in the message text? In MediaWiki, when using mw.message().parseDom(), that would be okay since it has a whitelist of tags and attributes, but we do not have that here. Messages are not editable by normal users, but you could easily get translator rights on translatewiki.net and sneak in evil translations, which would be imported and deployed to Wikimedia sites within a few days. I suggest that we only allow HTML in the parameters, which we provide from our code (or rather we should probably allow strings for plain text, and DOM/jQuery nodes for formatted text).

Are you planning to provide a different override for this in VE-MW? (If we disallow HTML, we probably shouldn't, to avoid inconsistent behavior.)

Patch-set: 2
Reviewer: Bartosz Dziewoński <417@e9e9afe9-4712-486d-8885-f54b72dd1951>