ve.dm.Document: Handle no valid offsets in #getNearestCursorOffset
A document should always have at least one valid structural offset in
it, but if somehow it appears that there aren't any,
ve.dm.Document#getNearestCursorOffset should not throw an exception.

It seems there is a bug in ve.dm.ElementLinearData#getRelativeOffset
in the code which is supposed to skip over nodes "whose children
should be ignored", such as figures (ve.dm.BlockImageNode). When given
an invalid offset at the beginning or end of such a node (inside it),
and asked to cursor outside of it, it will fail to find any valid
offsets and return -1.

This limitation is actually described in the code comments: "This
clearly won't work if you start inside such a node, but you shouldn't
be doing that to being with". Well, we do that, ever since we allowed
inline editing of figure captions (ve.dm.BlockImageCaptionNode) ;)

It should probably try the the other direction first, like it does
normally, and only fail if there are no valid offsets inside at all.

Add test cases for behavior of #getNearestCursorOffset near figures
and figcaptions.

Bug: T197651
Change-Id: I9ad2bc047ca353715b215a0ea4d4b77867752be2