Fix error in ve.dm.InternalList#sortGroupIndexes
Fixes the error "TypeError: Cannot read property 'getRange'
of undefined"

I found that the issue can be consistantly reproduced in case
a reference is reused with a key multiple times across paragraphs
and the reference content is not at the first use of reference.

Translating Atazanavir from English to Catalan is one example.
It is illustrated with a video in T196136.

Bug: T191011
Change-Id: I32d97fb4d3e6cf785e61d0fb5a8b17ca7edc046d