[WIP] ve.dm.Document: Remove incorrect handling for combining characters in #fixupInsertion
As a result of this code, we put two Unicode codepoints into a single
model position. As a result of that, model offsets and CE DOM offsets
do not match, and all hell breaks loose.

This code would have been neat if it actually applied NFC normalization
with String#normalize (to merge the base char and the combining char
into a single codepoint), but even then:
* String#normalize is not suppored by IE 11
* In some cases there is no single combining character (e.g. n̈ can
  only be represented as 'n' + combining diaeresis
* We have to be able to handle characters with arbitrary number of
  combining marks

I am guessing this was meant to prevent annotations that cover only
the base or the combining char (which causes weird rendering in
browsers), but that should be impossible to do from the editor (you
can't place a cursor from the UI between the two; it is possible if
you call setLinearSelection() though).

Bug: T85093
Change-Id: I6567b4290b0f9c850157a464d19bd6b8ad1982b8