[BROKEN] ve.ce.TextNode: Prevent browsers from collapsing trailing spaces in source mode
Change some spaces to NBSP to prevent the browser from collapsing
trailing spaces at the end of a line when rendering text. This uses
the same algorithm as CodeMirror (in function splitSpaces()) for
compatibility with syntax highlighting overlay.

BROKEN! The real and syntax-highlighted surface still drift away when
typing/backspacing space characters at end of line, because this code
doesn't run when typing. For this to work, we would have to re-render
the node on every keypress, and that would probably mess with IME
stuff. Maybe we could do it only for every keypress that inserts or
removes a space, but I couldn't figure out how to make that happen.

Restore some special handling for NBSP that was removed in
c6b338ad252c64a95aee1503d3083e5c7b3298bb. I'm not sure if the ve.ce.js
changes are necessary?

Bug: T188839
Change-Id: I86b480cb3a83e4ead9502cb1fbc33d09826b9efb