WIP proof of concept: QUnit tests in node
The problem with this strategy is there are three categories of code:
1. Code that runs in node only
2. Code that runs in both node and the browser
3. Code that runs in the browser only

However, some unit tests work best when code in categories 1 and 2
communicates with code in categories 2 and 3. The easiest way to achieve
this is to use the browser to run test code in all three categories.
Which makes this patchset redundant.

Change-Id: I879a67c7ab9864d11a36520e4b1622abfa4c4f2a