Update patch set 2
Patch Set 2:

I don't think npm audit should be part of the Grunt task. If you do want it to be part of CI and fail tests, then we already have a separate job that can be used for that.

Patch-set: 2
Reviewer: Legoktm <458@e9e9afe9-4712-486d-8885-f54b72dd1951>