ve.ce.Surface: Stop editor scrolling to selection constantly on iOS
iOS Safari has a "collapsible" address bar that becomes smaller when
scrolling down, and larger when scrolling up. This also resizes the
viewport, and the 'resize' handler here was triggering and scrolling
the selection into view, counteracting the user's actions.

This was not a problem before the great ditching of iOS hacks because
they resulted in the address bar always being in the expanded state.

This was also not a problem immediately after that because we missed
one of the hacks, which now caused all scrolling to selection to be
broken. Fixing that in I3e534a300e4b1286bb1a05c75834a93fc3cf5dec
exposed this bug.

Bug: T218635
Change-Id: I02ec12d8dddacdd0efa88a6094d77ad8d8236fbd