Update patch set 1
Patch Set 1:

We currently use it like this to analyse hard-to-reproduce IME issues on live sites. More than once before we've solved an issue by asking the reporter (whose IME we don't have access to) to open the browser console on Wikipedia, type a command and then send us the dump.

Back when this was written, we decided it wasn't at that time worth the coding effort of wrapping this all in a bookmarklet to get the same effect.

Patch-set: 1
Reviewer: Divec <839@e9e9afe9-4712-486d-8885-f54b72dd1951>