Make null-selection-on-blur optional
If nullSelectionOnBlur is set to false, we just deactivate
the selection when the document is blurred, but don't set
a null selection.

Change the treatment of deactivated selections (linear,
table & focusable nodes) to grey to make it clear the
native selection is not present.

Remove blur/focus event listeners from DesktopContext, these
have not been necessary for many years, and in the new mode
they unnecessarily hide the context on deactivate.

Bonus: Fix an issue where the first click blurring a
document is ignored, as the call to #deactivate put the
native cursor back in the $pasteTarget.

Bug: T208826
Bug: T219813
Bug: T219829
Change-Id: I2c464ec2bbf799cbe1d9bfed1aad05c5a1bfb094