Update patch set 1
Patch Set 1:

Hmm. I think we do, although now only in the case where the selection is collapsed on mobile, for insertion annotations. Currently the only dialog that produces insertion annotations is Language, so it is an edge case, but I think we have to chose between activating in the user event, or after the dialog has performed the action, unless we rewrite the window manager.

Patch-set: 1
Reviewer: Esanders <658@e9e9afe9-4712-486d-8885-f54b72dd1951>