Update patch set 6
Patch Set 6: Code-Review+1

To confirm, the abstract and standalone version of the Storage class, can be omitted from the MW-VE bundle? Sorry if that's obvious, but couldn't tell right away. If not for some reason, then might be easier for VE to always use its own given the MW version isn't doing anything extra (not even a prefix).

LGTM.

One note about localStorage - For storing anything personal, generally recommend against because it can persist indefinitely. It has no TTL, is not bound to the MW login session, and also not bound to the tab session, or browser process. It survives pretty much anything, which makes it potentially unsuitable for storing something the user entered on a page. I don't know how much (if anything) we care about the use case of shared/public computers, but to my knowledge we've currently (intentionally or not) always accommodated it. So sessionStorage would be more suitable for that. Although it has the obvious downside of being counter-intuitive to non-techsavvy users. Specifically, sessionStorage is pretty good when it comes to using a tab to navigate around, and backwards etc, even re-opening from "last closed tabs" etc, and even if you've visited other origins in-between. But as soon as you try to go to the "same place" as a user from another tab open to the same website, it's not there (each "tab" life cycle has its own sessionStorage). So yeah.. anyway.

Patch-set: 6
Reviewer: Krinkle <34@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=+1