WindowAction: Refactor surface re-activation logic once again
The main change here is to move the `.isCollapsed()` check inside
the `windowManager.once( 'closing', function () { … } )` callback.
Previously we were incorrectly checking it when the dialog was
being opened, rather than when it was being closed.

But doing so resulted in the if…else structure making no sense.
I restructured it and I hope it makes some sense now.

Bug: T220633
Change-Id: I2d4a3649879ecc4d95d7daeda4e040dbf18c5561