Move logic for new mobile contexts to mouse down handler
This prevents the keyboard from showing while we decide
if we are going to show a context.

Refactor out some logic from LinearContext so we can
calculate if a context is about to be needed in advance.

Also a small transition to smooth some of the clashes
with the virtual keyboard.

Bug: T226073
Change-Id: Ia439fbd52631e2265b7f4daebab873af1b3cae49