Fix TableNode unit test getOffsetFromCoords failure on Firefox
Our ve.ce.TableNode unit tests use a fake click event, with dummy x,y coordinates that don't
meaningfully relate to the test surface. On Chromium and on Firefox non-headless, it just so
happens that an error is raised causing the method to return -1, upon which the test succeeds.
But on Firefox headless it returns 0 which leads to a test failure.

Therefore, use a fake getOffsetFromCoords instead.

Bug: T228767
Change-Id: Ic1a095b81f18d208d3afd79810809e138df5eb89