Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> we should think about the use cases where we want the clear button, otherwise we'll end up adding it everywhere

My initial thought was that anything with autocomplete ("search", in practice) should definitely have it, just because it's going to be fairly easy to accidentally drop a long string in and then want to change the entire thing.

As a guiding principle, I'd think any field where a _likely_ change is "I want to replace this entire thing". For URLs like this, it seems very plausible that the most common editing pattern is really going to be clearing it and then pasting the new version in, even if it's technically a small 1 character change somewhere in the URL.

We might want to take this elsewhere and rope Jess in, for a design perspective.

Patch-set: 2