Update patch set 4
Patch Set 4:

Mostly passes now. The left-over issues are genuine, in so far that they are stuff that explicitly is trying to copy text from JSDuck indexes, @inheritdoc - which can't anymore naturally. Should probably just be removed and have their own descriptions. This is probably good so that one doesn't have to go to another repo or upstream lib file to understand the VE code. Could still use a @see <url> (instead of @inheritdoc) to point to upstream for more details (if there are any).

Patch-set: 4
CC: Krinkle <34@e9e9afe9-4712-486d-8885-f54b72dd1951>