Update patch set 4
Patch Set 4:

> missing whitespace regex fix on PS4
Oops, added.

I'd also be interested in feedback on whether you feel the shouldAbandon method there has a more elegant alternative. I felt that putting it in the Action made sense, since different actions might have different rules for when to abandon... but the number-of-matches parameter feels a bit awkward to me.

Patch-set: 4